+++ Installer/Assets/Unity Collections Specialized Installer/Installer.cs
@@ -16,7 +16,7 @@ namespace Unity.Collections.Specialized.Installer
     public static partial class Installer
     {
         public const string PackageId = "com.saesentsessis.unity-collections-specialized";
-        public const string Version = "0.1.1";
+        public const string Version = "0.2.0";
 
         static Installer()
         {